Move the Client code into the queries module
This type is used extensively in the SDK, and thus it should be
considered under the SDK banner
