+++ apps/src/lib/cli/api.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use namada::ledger::queries::Client;
+use namada::sdk::queries::Client;
 use namada::sdk::rpc::wait_until_node_is_synched;
 use namada::tendermint_rpc::HttpClient;
 use namada::types::control_flow::Halt;
