+++ shared/src/sdk/mod.rs
@@ -6,3 +6,5 @@ pub mod masp;
 pub mod signing;
 #[allow(clippy::result_large_err)]
 pub mod tx;
+
+pub mod queries;
